fixes #29 - look for common sub dirs for discovery
With this change relcool looks for the common sub directories to add
to the search. Those directories are now, relative to the project dir,
`apps`, `lib`, `ebin`, and the release output dir.

These automatic subdirs can be disabled by setting
`disable_project_subdirs` to `true` in the config.

The release output dir is also added to the discovery process. This
can be disabled by setting `disable_discover_release_output` to true.