Support upgrade and downgrade between versions
The "bin/RELNAME install" command now handles upgrading and downgrading.
It unpacks and installs new versions as needed, or just sets existing
releases as current/permanent as needed.

I renamed from "bin/RELNAME upgrade" because it handles downgrades too.
"upgrade" and "downgrade" are now synonyms for "install"

Additionally, when installing a new to-be-unpacked version, you can say:

  bin/RELNAME install 0.1.1/beat

OR just:

  bin/RELNAME install 0.1.1

in which case the /RELNAME part is added automatically. This is to keep
the version spec for downgrades the same as for new upgrades.

This is fully backwards compatible with the existing "upgrade" command.