Add support for explicit erts path
Add support for 'include_erts' to be defined as a path to an external erlang
release.  This will allow for relx to include cross-compiled erts builds.

include_erts can now be one of three values: true (default), false, or
"/path/to/erlang".

The symantics do not change unless it is set to a string in which case
it only overrides the default lookup "code:root_dir()" with the given
path.