Added the 'escript' command to the extended release script to permit escripts to run in the release's environment without having to break release encapsulation.
Example use:

$ _rel/myrelease/bin/myrelease escript priv/test.escript arg1 arg2 arg3

The argument to 'escript' is relative to the release's root.

The script is started with the 'escript' binary in the release's erts
release; has a current working directory of the release's root directory;
and additionally is started with visibility of an environment variable
called 'RELEASE_ROOT_DIR' which contains the absolute path of the
release's root directory.