Remove bashisms from run scripts
This commit removes the bashisms in `bin` and `extended_bin`.
Both of these scripts used `local` variables which are a bash
addition and aren't supported on Solaris/SmartOS /bin/sh. To keep
the local intention of the variables, they were renamed from $var to
$l_var.

In addition, `extended_bin` used `kill -SIGNAL $PID` which is also
not in Solaris `kill`.