Include src/ in tarballs unless specified not to
By default, the `include_src` value is `true`, yet Relx disregards the
value entirely when generating tarballs.

This patch makes it so the value is used and respected when generating
tarballs. Given the default value is `true`, this patch changes the
default behaviour of relx to always include source files in tarballs,
considering its omission a bug rather than a feature.

Tests are added, verifying the impact on ERTS-included files.