Don't rewrite app files if contents are the same
When comparing .app files to check if a rewrite is
necessary do so by comparing erlang terms instead
of bytes, the original .app might contain comments
for example and cause unnecessary rewrites.
This proves a problem when dealing with erts .app
files that might be read-only at the origin and
thus cause a failed release on relx that is 
uncalled for since there might have been no changes
at all.