Maintain no_dot_erlang.boot file in releases
Had a problem in OTP-21 (master) where escript calls in nodetool would
fail since the file is not around and is being used by default there.

The patch works by copying its equivalent file, which I belive to be
just the start_clean file (their .rel.src are identical), into the same
directories as we do for start_clean.

Internal errors are renamed to be made neutral.