Windows: support remote_console, fix attach
On Linux, 'attach' uses named pipes, and 'remote_console' uses -remsh.
The latter is usually deemed better since named pipes require a call to
fsync on every line written.

On Windows, no named pipes are available so attach uses -remsh directly.
Historically, remote_console was added to linux *after* attach, but no
alias was added for it on windows. Since there's a predominance of
tutorials using linux-likes, remote_console is widely documented as the
way to go, and is unavailable on windows.

This is hella confusing.

So to work around that, this patch adds an alias for 'attach' on windows
to be 'remote_console', bridging the gap.

Also the functionality was flat out broken because it would not use a
node hostname when connecting out. Since the latest release added that
functionality, this patch also fixes attach to work in the first place.