Add support for overriding start script file type on Windows
This is necessary when using cygwin or msys2 shells for doing
release management and application startup.

The feature is implemented by introducing a `start_script_type`
configuration option, which can take values: `unix` or `win32`.
The default value is determined by a call to `os:type/0`.

When `unix` is provided in the Windows environment, the start
script will be generated that is suitable for running by a
unix shell rather than Windows `cmd`.