Implement support for cygwin/msys2 by extended script
With this change a user can use cygwin or msys2 bash shells
on Windows to manage release startup.

For running console two new command-line parameters are provided
in addition to the existing ones:
  1. wconsole
  2. wconsole_clean

They will start `werl.exe` instead of `erl.exe`, and run that as
a background task, so it won't block the current shell.