Implement systemd-like config support for sourcing hook files
1. A new variation of the `extended_start_script_hooks` option
is added that can define a source directory containing hook
scripts: `{directory, HookScriptsDir::string()}`. When defined,
the scripts will be copied to a sub-folder in the release at
the level of the extended start script.  That subfolder
will have the same name as the basename of the extended start
script (without the release version number and extension), and
will have a `.d` suffix.

2. That subdirectory can contain `pre-/post-` `start/stop/install`
scripts with by saving them in the startup script's release directory
that has a `.d` suffix. So, if the extended script's name is `myapp.sh`,
then the hook directory can be:
```
bin
  +-- myapp   <-- Extended start script
  +-- myapp.d
    +-- pre-start.NN
    +-- post-start.NN
    +-- pre-stop.NN
    +-- post-stop.NN
    +-- pre-install.NN
    +-- post-install.NN
```
The hook extension files can have any suffix (such as `.NN` shown above).
The files in this directory will be sourced only if the corresponding
`PRE_*` or `POST_*` variables (populated by respective `pre_*` and `post_*`
hook configuration options) don't contain script names that override
this behavior.