add log_funs to state and use for new log layer
Logging is now done through marcos in rlx_log.hrl. The functions
called in rlx_log module check if the relx state has a log fun
for the log level being using. If there are no log_funs set in
the state then it writes to stdout with io:format. If there is
then it calls that function with Msg and Args, or if the atom
no_level_fun is returned the log is ignored.

This way rebar3 can set a function for each log level and relx
will call rebar3's log functions.