overlays: base is_directory only on the overlay To variable
copying a directory can result in a new directory exist on the
next run, causing a different result for the same overlay. So
this change removes the check to see if To in an overlay copy
is a directory on disk and only checks if it ends in `/` to
signal it is a directory.