new defaults and new options for systools on sasl >= 3.5
use new `extra_files` option when creating a tarball on sasl 3.5+ so
that the tarball does not have to be unpacked and recreated.

excluding erts no longer defaults to excluding system libs like stdlib

src is excluded from the release by default

src_tests is enabled by default which checks if beam files are older
than the src file that created them and warns if so.