merge use of nodetool and erl_call depending on ERTS version
this patch updates the extended start script to check for a new
enough erl_call and use this for rpc and eval calls if available.
The release will continue to bundle in nodetool as a fallback
even for releases built for OTP-23+ since it is simpler to do the
check at runtime.

the start script rpcterms option is removed and the rpc option
has changed so that both erl_call and nodetool work the same.
Arguments to function call must be a quoted list ending with a
period, like `bin/foo rpc bar waz '[1,2,3].'` will call the
function `bar:waz(1,2,3)` on the running node.