remove remote_console test that was never working
this test was supposed to be checking that the remote console was
attached but the remsh was actually failing and the test "passed"
because the node list would contain the node making the call to
get the nodes list. It should have been checking for 2 not 1. But
remote_console will fail always when run like this because it can't
run in the background.