move create_RELEASES function to relx
The version in sasl's release_handler module required use of
set_cwd to get a result where the application list had the apps
as relative to the root of the release -- "./lib/..." -- instead
of requiring a newer version of OTP (after getting a new version
into OTP) having this small function added to relx until we only
support versions with the new release_handler module is not a
burden and makes things simpler.

Part of why set_cwd is bad is it makes parallel release building
not an option because of conflicting changes to cwd.