Preserve the order of goals in rel file
According to the systools documentation:

> The applications are sorted according to the dependencies between the applications. Where there are no dependencies, the order in the .rel file is kept.

This allows a certain level of control of the application start
sequence, which, unfortunately, is not available if one is to generate
releases using relx, due to the use of an unordered container (map) to
store release app goals.

Say, you have two applications, without any dependencies between them,
that you'd like to include in the same release. The release spec would
then look like: `{release, {my_release, "0.0.1"}, [b_app, a_app]}`.
Expectation is that, since there are no dependencies between them, b_app
would be first in the boot script, which is not the case due to the way
`maps:to_list/0` _currently_ works.

This change adds an extra field to the `parsed_goal()` map, which stores
the index of the goal in the original release specification. Goals order
is then restored according to this index value prior to generating the
.rel file. No other functionality is affected.