updates to start script for otp 23.1
If erts 11.1 or above is used then `undefined` is to be used for
the longname in remsh. This results in a generated node name.

When ERL_DIST_PORT is set and no start_epmd argument is found in
the vm args file and the erts is 11.1 or newer then add start_epmd
false to the dist args automatically.

This means on the latest Erlang the user needs to only set
ERL_DIST_PORT to get a static port for the node and disable epmd.