Stores the warnings from xref as part of the rlx_state instead of printing them.
Summary:
Before, the warning from xref were printed to the console,
Wthis this commit, the warnings are stored in the rlx_state, and the required modifications are carried to make sure the entries command of the module all returned rlx_state.

This transfer the responsability to deal with them to rebar3.

Notably, it allows rebar3 to call code from the rebar_prv_xref plugin to disregard warnings that should be ignored by xref (-ignore_xref tag in the modules or in rebar.config).

Test Plan: rebar3 ct && rebar3 dialyzer
+ on a custom project, with custom rebar3 (see pull request... )
rebar3 release

Reviewers:

Subscribers:

Tasks:

Tags: