State extended with a filter method to filter xref_warnings
We extend RELX state with a method that is used to filter XREF warnings.

This method is by default just the identity function.

This is desgined in such a way that rebar3 can pass it the method it currently relies on to deal with the ignore_xref attributes in the future, while offering no regression with current rebar3 versions.