add support for optional_applications in otp24
Optional applications may or may not also be in the applications
list. They are also only optional for the app that lists them
as optional, unlike an excluded application which is global. So
what is optional for app A may be required by app B and thus
cause a failure if it is not found when building the release.