fix: Write os-var replaced .src file to temp file, and rename
The replace_os_vars function is called for any command, not only those that boot the release. If you invoke <release> status in a readiness check, that will write the file again.

This might lead to race conditions, where the release starts up, but the health check will re-render the sys.config and write it halfway while the boot gets around to read it.