Force -mode(interpret) in templated escripts
Starting with OTP-27, the escript files are compiled by default instead
of running in interpreted mode. Unfortunately for Relx, this causes
issues for releases when the application that has been shipped does not
bundle the compiler application, and errors come out of it.

For full compatibility and better usage for releases, force files to be
interpreted at all times, removing a hidden dependency on the compiler
being present.

Fixes https://github.com/erlware/relx/issues/940