extended_bin_windows: Use werl.exe as symlink for OTP >= 26
As per the OTP-26 release notes:

"The TTY/terminal subsystem has been rewritten. Windows users will
notice that erl.exe has the same functionality as a normal Unix shell
and that werl.exe is just a symlink to erl.exe. This makes the Windows
Erlang terminal experience identical to that of Unix."

Due to this, the werl.exe was missing for OTP >= 26 on Windows releases.

## Solution

- Dynamically set `werl.exe` to use `erl.exe` when OTP ≥ 26 is detected
- Add version check using `erlang:system_info(otp_release)`
- Use quoted string comparison (`geq "26"`) for safe version checks